+++ src/util/udp_client.cc
@@ -12,11 +12,11 @@
 
 namespace ERpc {
 
-UDPClient::UDPClient(
-    const char *remote_addr, size_t remote_port, double drop_prob)
+UDPClient::UDPClient(const char *remote_addr, uint16_t remote_port,
+                     double drop_prob)
     : remote_addr(remote_addr), remote_port(remote_port), drop_prob(drop_prob) {
   char decimal_port[16];
-  snprintf(decimal_port, sizeof(decimal_port), "%zu", remote_port);
+  snprintf(decimal_port, sizeof(decimal_port), "%u", remote_port);
   decimal_port[sizeof(decimal_port) / sizeof(decimal_port[0]) - 1] = '\0';
 
   struct addrinfo hints;
